Leverage Cisco Bootstrap Day0 Configuration
Leverage the Cisco bootstrap day0 configuration method. We use
a configuration file for each node, generate a small ISO file
and mount that as a disk into the VM when it boots. When the IOS
image starts up, it reads the configuration in the ISO and does
the initial configuration for us.

By doing this we don't need to use an awkward telnet method
which wasn't working with the virtual machine start up due
to missing prompts on connection.